add makefile

+++ server/include/db_folder.h
@@ -7,6 +7,6 @@
 db_errror_code db_folder_create_root(char *username);
 
 // Tìm id của thư mục gốc của user có thư mục gốc
-db_errror_code db_folder_find_root(int user_id, int *root_id); 
+db_errror_code db_folder_find_root(int user_id, long *root_id); 
 
 #endif
